requirements.txt
pycryptodome
requests
Flask==1.0.2
